refactor: 공용 ui 컴포넌트 분리 - eslint 타입 오류 수정
